#fix unsigned型とsigned型の暗黙の型変換を主うすぃ #disable C4642警告の抑制を無効化 #add SuzukiAlgoのsearch_depthに[[maybe_unused]]を追加

+++ MainProject/KosenProcon30/CommunicationData.hpp
@@ -24,7 +24,7 @@ namespace Procon30 {
 		String host;
 
 		//‡‚Ì”
-		size_t matchNum;
+		int32 matchNum;
 
 		// gameID‚Ì•ÏŠ·ƒe[ƒuƒ‹
 		std::unordered_map<int32, int32> matchesConversionTable;
